+++ src/plonemeeting/portal/core/browser/homepage.py
@@ -2,12 +2,17 @@
 from plone import api
 from plone.memoize import ram
 from plone.protect.interfaces import IDisableCSRFProtection
+from plone.restapi.interfaces import ISerializeToJson
 from plonemeeting.portal.core import logger
 from plonemeeting.portal.core.config import DEMO_INSTITUTION_IDS
 from plonemeeting.portal.core.config import LOCATIONS_API_URL
 from plonemeeting.portal.core.config import REGION_INS_CODE
+from plonemeeting.portal.core.interfaces import IInstitutionSerializeToJson
 from Products.Five.browser import BrowserView
+from zope.component import getUtility
+from zope.component import queryMultiAdapter
 from zope.interface import alsoProvides
+from zope.schema.interfaces import IVocabularyFactory
 
 import json
 import requests
@@ -38,11 +43,19 @@ class HomepageView(BrowserView):
         for brain in brains:
             if brain.id not in DEMO_INSTITUTION_IDS:
                 institution = brain.getObject()
-                institutions[brain.id] = {
-                    "title": institution.Title(),
-                    "URL": institution.absolute_url(),
-                }
+                serializer = queryMultiAdapter((institution, self.request), IInstitutionSerializeToJson)
+                institutions[brain.id] = serializer(fieldnames=["title", "institution_type"])
         return json.dumps(institutions)
+
+    @ram.cache(institutions_cachekey)
+    def get_json_institution_type_vocabulary(self):
+        """
+        Get institution_type vocabulary from this portal and serialize it in JSON
+        """
+        factory = getUtility(IVocabularyFactory, 'plonemeeting.portal.vocabularies.institution_types')
+        vocabulary = factory(self.context)
+        serializer = queryMultiAdapter((vocabulary, self.request), ISerializeToJson)
+        return json.dumps(serializer("institution_type"))
 
     def get_faq_items(self):
         """